telegram_file_service/pyproject.toml
[tool.poetry]
name = "telegram-file-service"
version = "0.1.0"
description = "Telegram bot + notify worker (webhooks) for audio summarise pipeline"
authors = ["Anton"]
packages = [{ include = "telegram_file_service", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.7"
fastapi = "^0.115"
uvicorn = "^0.30"
httpx = "^0.27"
python-dotenv = "^1.0"
aiokafka = "^0.10"
boto3 = "^1.34"

[tool.poetry.scripts]
tg-bot = "telegram_file_service.app:run"  
notify = "telegram_file_service.notify.worker:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"